add login condition
